# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api swagger collection for DATING APPS API.",
        "title": "DATING APPS API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Abdul Halim",
            "url": "https://github.com/h4lim"
        },
        "version": "1.0"
    },
    "paths": {
        "/login": {
            "post": {
                "description": "Login API.",
                "tags": [
                    "Dating Apps API"
                ],
                "summary": "Login API",
                "parameters": [
                    {
                        "type": "string",
                        "default": "application/json",
                        "description": "Must application/json",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "EN",
                        "description": "Must be EN or ID",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RequestLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.ResponseLogin"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.GinResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.GinResponse"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "Signup API.",
                "tags": [
                    "Dating Apps API"
                ],
                "summary": "Signup API",
                "parameters": [
                    {
                        "type": "string",
                        "default": "application/json",
                        "description": "Must application/json",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "EN",
                        "description": "Must be EN or ID",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RequestSignup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.GinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.GinResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.GinResponse"
                        }
                    }
                }
            }
        },
        "/swipe": {
            "post": {
                "description": "Swipe API.",
                "tags": [
                    "Dating Apps API"
                ],
                "summary": "Swipe API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token get from login",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "application/json",
                        "description": "Must application/json",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "EN",
                        "description": "Must be EN or ID",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RequestSwipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.GinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.GinResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.GinResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.GinResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.GinResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
        "types.RequestLogin": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.RequestSignup": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "confirm_password": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.RequestSwipe": {
            "type": "object",
            "properties": {
                "is_first_view": {
                    "type": "boolean"
                },
                "profile_id": {
                    "type": "integer"
                },
                "right_swipe": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.ResponseLogin": {
            "type": "object",
            "properties": {
                "expired": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.ResponseSwipe": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "premium_package": {
                    "type": "string"
                },
                "profile_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}